Use alloca() instead of heap allocation to store vfork's jmp_bufs.
Previously we were keeping a linked list of objects storing the jmp_buf
structures used to record (via setjmp()) the context of vfork() calls (to be
restored with longjmp()).

Now, we allocate space on the stack, and pass a pointer to it to _vfork_start(),
a wrapper for the vfork_start() system call, which stores them in members of the
process' LongJumper base.  After exec or exit, the LongJump() method will call
the function and pass it the buffer address and the value to return from
setjmp().  (This is the same information passed to longjmp() itself, but the
callback takes a (jump_buf*) instead of a (jmp_buf).)

Since vfork_start() is called from a library function instead of directly by the
caller of vfork() (which is inlined), it's now one stack frame further away from
the kernel code recording the frame pointer.