Implement utimensat() and futimens(), and remove custom routine.
Remove all traces of futimesat_k().

Modify FSTree::SetTimes() and SetFileTimes() to take an array of two timespecs
(like the POSIX routines) instead of four pointers to timevals.

Define UTIME_NOW and UTIME_OMIT as in Linux.

Define UTIME_ARCHIVE to 0x40000000, or (1 << 30).  When set in tv_nsec, atime
is 'archive' or backup time.

Define futimens() as a wrapper around utimensat(), as on Linux.

Split utime() into its own unit and reimplement in terms of utimensat().

In jsync, copy the modification date as a timespec if available, and set via
futimens(), or futimes() on OS X.