Set gCurrentProcess to NULL before process thread terminates.
We were letting the process' thread terminate without zeroing gCurrentProcess.
Usually, this would become moot as soon as the next process switched in.
But in the case of ProcessList::Killall(), the reaper thread ran immediately
after, and when a bus error occurred, Genie attempted to deliver a signal to a
process that no longer existed.