Release the controlling terminal and current directory on process termination.
The process holds a shared_ptr to a process group, which holds a shared_ptr to a
session, holding a shared_ptr to a controlling terminal filehandle.
Release the latter indirectly by way of the former before terminating the thread
so that files (e.g. the console window) are closed on this thread rather than
the reaper.

Similarly, release the current directory, which may be an additional reference
to the same terminal (as with /sys/window/abcd3210/tty).