Refactor CloseUserWindow() into multiple smaller functions.
CloseUserWindow() does two different things: disconnecting a window's terminal,
and destroying the window (uninstalling its views and disposing the WindowRef).

When the user requests window closure, (e.g. by choosing Close or clicking the
close box), this should disconnect the terminal only (and allow the foreground
process group to actually close the window, usually by terminating and closing
the last file descriptor to it), unless the window is not associated with a
terminal, in which case the window should be closed immediately.

If the last file descriptor to a window is closed when a terminal is still
present, then both operations must take place.