Don't zero st_size in FSTree::Stat().
If st_zero is left as is (initialized to -1), then fstat() and fstatat() will
call GetEOF() on the filehandle or file to fill it in.  In addition to regular
files, we need this for symbolic links, so an lstat() on /dev/fd/0 opened to a
property file in non-binary mode yields the size of the non-binary data.