Make IOHandle::GetFile() non-const.
This change is needed so IOHandle::GetFile() can pass shared_from_this() to
a new FSTree_IOHandle.  GetFile() doesn't actually change anything, but it
creates (and returns) an object containing a non-const reference to itself.

This has some collateral ripple effects for filehandles which are closely tied
to their underlying files.  In TextEdit_text_Handle, for example, the following
member functions form a call chain:  GetEOF -> String -> ViewKey -> GetFile.
For GetFile() to be non-const, so must ViewKey() be, and then String(), etc.
This ripples back to RegularFileHandle via GetEOF() and back down to all classes
derived from it (and overriding GetEOF()).