Always sync TE scrap on edit.
Traditionally, TE scrap is synchonized on a layer switch, i.e. suspend/resume.
Aside from the risk of losing the clipboard if the app exits unexpectedly
(either by crashing or even merely calling ExitToShell()), Basilisk II and
SheepShaver propogate the clipboard between the OS X host and the traditional
Mac OS guest, and sync-on-switch just doesn't play well with this feature.

Carbon introduced sync-on-edit:  The TE scrap is flushed immediately (on TECut
or TECopy) and updated lazily -- not until it's needed (if ever) -- on TEPaste.
Now we adopt this protocol for classic TE scrap as well.