Add /sys/window/REF/{accept,cancel}.
These files are symbolic links if they exist at all.  Intended practice is to
link them to the 'click' devices of OK and Cancel buttons.

A subsequent patch will these files to be written to in response to the
appropriate user input, e.g. 'cancel' when the user hits Escape.