Store local and remote mod dates in the base file's metadata.
It's not safe to copy modification dates at all.  Even the initial copy might be
following an accidental deletion of a file that was compiled since the source
was modified.

Our solution is to copy the modification dates fo the local and remote files to
metadata slots on the baseline file.  The base file's mod date is copied from
the local file, and the backup/archive/checkpoint date is set to the remote
file's mod date.

If the files have non-matching dates but matching contents, store the new dates.
Don't do anything immediately after rewriting an out-of-date file, but after
writing the baseline file, store the new dates.