Guard out units requiring Mach-O with #ifdef __MACH__ covering the primary file.
This is the easiest way to support limited-availability APIs when available.
That is, by allowing the primary file to be compiled in all cases, and guarding
out code when it's unsupported for that platform.  This works whenever support
can be determined by testing preprocessor macros.

While we could have used A-line platform config candidates in this case (mach-o
vs. !mach-o), it won't help for APIs introduced in later OS versions.