Catch errors as OSStatus instead of ErrorCodes.
Metrowerks C++ 2.4 doesn't distinguish between types which differ only with
respect to a numeric template parameter for the purpose of catching exceptions,
so an attempt to catch a specific ErrorCode is in error.  The simplest
workaround is to always catch the ErrorClass type (here, N::OSStatus) and check
explicitly for the right error code and throw otherwise.

Also, check for OSStatus -10661 launching ToolServer on Carbon, not just Mach.