NitrogenLib:  Search among Mac OS headers before among the source files.
A-line automatically appends the directory containing the currently compiling
primary file to the search path.  Since this takes priority over system-supplied
search paths (e.g. CIncludes), an include of <Movies.h> finds the local
Nitrogen/Movies.h instead of the one in CIncudes.[1]

This hasn't been a problem before because the local directory is only added for
compiled primary files, not precompiled headers.

To counter this, we add the CIncludes project as a dependency of NitrogenLib.
This ensures that including Mac OS headers works as expected.

A future solution might make the local directory search optional per project.

[1] Why this only happened with Movies.h is a complete mystery to me.