Kerosene/Library:  Fix buffer overflow in gethostname().
If the nodename is exactly the length of the buffer (in which case the trailing
NUL doesn't fit), the length check succeeded anyway and wrote past the end.

For simplicity, add one to the nodename length up front.
For brevity, replace std::copy() call with memcpy().

Also, remove unused include of assert.h.