backtrace:  Remove now-unused set_stack_bottom_limit().
This mainly existed to avoid falling off the end of the stack in Thread Manager
threads on PPC, but now we patch the stack to end in a null frame pointer.

Secondarily it was used to trim a Genie program's stack crawl down to main(),
but this is easily accomplished by comparing frame pointers after the fact, with
the added bonus of not depending on a global variable.