recall:  Fix buffer overflow in make_report_for_call().
When we added the frame pointer to the output we forgot to reserve space for it.

In practice, frame offset numbers are limited to two digits, so we're only one
byte over, and since the allocated size is an odd number (33 bytes) the extra
byte is left as padding anyway.  But it's still an error.