Genie/FS:  Rename a volume on write to /sys/mac/vol/N/name.
We use HRename() for this purpose, which does exactly what we want.

Another option was the pair of PBHGetVInfo() and PBSetVInfo(), but aside from
the disadvantages of requiring two calls instead of one and generating larger
code, this has a curious failure mode in the case of unrenameable volumes (e.g.
AppleShare mounts):
The new name is not displayed in the Finder or accessible via the drive index
(as in /Volumes), but it *does* 'stick' when accessed by vRefNum, so reading the
'name' file yields the new name.  Smells like a bug in either the File Manager
or the AppleShare client, but in any case HRename() doesn't have this problem.