Use sig{jmp_buf,setjmp(),longjmp()} instead of standard C variants with vfork().
Also, rename _vfork_start() to ensure only recompiled sources link successfully.

sigsetjmp() and siglongjmp() have been modified to save and restore the backlink
on PowerPC in case a subsequent alloca() caller clobbers it.

This solution is superior to an alternate workaround, to pad the allocation size
so nothing gets clobbered.  That depends on knowing the offset from top of stack
to alloca() result, as well as how much space lies unused at the end of jmp_buf,
though the latter could be safely ignored.  Though not yet tested for such, this
solution might also work for Altivec.

Also, saving the signal mask isn't a bad idea either.

[The following message documents the problem and the previously used workaround]

Pad the jmp_buf size passed to alloca() for MWCPPC to prevent munging the stack.

Metrowerks' __alloca() built into MWCPPC extends the stack by the size requested
(aligned to 8 bytes).  Since the top of the stack contains a frame header (which
includes the previous SP, at least), the new storage must be offset from the top
and therefore encroaches the previous location of the frame header, (and notably
clobbers the previous location of the saved SP).

A vfork's corresponding longjmp() restores the SP that was in effect at the call
to setjmp(), whose backlink could have been clobbered if alloca() was called (as
a subsequent vfork() would do).  In this failure mode, when the function cleanup
code pops the stack, the SP is set to the clobbered backlink (an arbitrary value
from the top word of the saved fp30).  (Hilarity ensues.)

Our evasive maneuver is to pad the allocation.  So far, the offset of alloca()'s
result from the new top of stack has been observed to be 56 bytes.  However, the
the internally used struct occupies only 248 bytes, while the publically defined
jmp_buf type spans 280 bytes, a difference of 32 bytes.  So only 24 bytes lie in
the overlapping area, and we pad the allocation size by that amount.