Lamp headers:  Include <ctime> from time.h for CLOCKS_PER_SEC definition.
MSL defines (by default) CLOCKS_PER_SEC to 60 for Mac OS, although it defines
clock() in terms of TickCount(), which prior to OS X counts ticks slightly more
frequently than 60 per second.  Calculations involving large lengths of time
derived variously from clock() and other sources (such as times()) are likely to
be wrong.

MSL also provides the option of defining clock_t as a double, in which case it
defines CLOCKS_PER_SEC to 1000000 and bases clock() on Microseconds().

A third option to consider is using Microseconds() as above, but with uint64_t
instead of double.