Add functions {read,write}_{some,all}().  Initial check-in.
The functions read_some() and write_some() are, to xread() and xwrite() (as they
appear in git, at least) are identical in purpose and similar in function.

Aside from the obvious proviso of being poseven functions (and consequently
throwing exceptions instead of returning errors), the key difference is that
read_some() and write_some() (and by extension, read_all() and write_all()) pay
no particular attention to EAGAIN.  Whereas on EINTR these functions all DWIM
and retry the operation, EAGAIN (a) never occurs with blocking file descriptors,
and (b) otherwise indicates that the operation *will fail again* until a certain
event occurs (as opposed to EINTR indicating that the operation failed *because
an event occurred*).  Since the lack of available data indicated by EAGAIN may
persist for some time, an otherwise unbounded loop retrying the operation is
broken with respect to performance.  The correct solution is either to use
blocking mode, or (if other users of the file descriptor rely on it remaining
non-blocking) to call select() or similar between attempts.

The functions read_all() and write_all() are analogously similar to git's
read_in_full() and write_in_full().