Use DefaultValue_Traits so the default value can be customized.
By default, DefaultValue_Traits< Resource >::DefaultValue() returns exactly what
it replaces -- a default-constructed object.  But clients can specialize the
template to have it return something else, like fd_t( -1 ).

Since the resource will be constructed with copy construction instead of default
construction, this will even work for structs (such as AEDesc) which currently
are unintialized and require the use of SeizedValuesAreLive and the extra member
variable it uses.