tlsrvr:  Use plain FSSpec instead of Shared< FSSpec > for temp files.
The impetus here is removing Owned< FSSpec > from Nitrogen, but this is worthy
in any case.  A thrown exception will prevent deletion of the temp files, but
this could be argued as a good thing.  Also, the more likely scenario is a fatal
signal, which would have prevented the static objects' destructors from running
anyway.

This cuts 2-3% off the executable size.