Nitrogen:  Remove support for Owned filesystem objects.
Files are owned by the filesystem.  Of particular relevance is that they can be
enumerated, and therefore can't be leaked as such -- so returning Owned objects
from file creation routines (which upon destruction will 'clean up' their files
by way of deleting them) is uncalled for.

It can certainly be useful to use Owned on files, but (a) it's up to the user to
elect this behavior, and (b) there are other options than simple item deletion,
like recursive deletion (one of which it's again the user's job to select).