Specify std::min() as std::min< size_t >, since off_t may be larger than size_t.
In 32-bit OS X, size_t is a 32-bit type but off_t is a 64-bit type, so min's
type parameter can't be inferred.

Note that errors remain in the case where offset + n_bytes exceeds size_t( -1 ),
regardless of the width of off_t, and 64-bit file access is not supported.