Add new 'debug' project, with configs for systems with sane ld and for Unix.
Unix ld resolves references only using later libraries, not preceding ones.
A-line tries to accommodate this limitation by basing the link order on project
dependencies.  This works as long as a project (or one of its own prerequisites)
in fact declares a dependency on the required library -- if a dependent project
specifies the dependency, then the Metrowerks linker will handle it just fine
but ld will complain.

So to appease the 'special' systems, debug is an alias for debug-stdio+abort,
which is always appropriate for Unix.  For the others, it's just debug-headers.