text_input::feed:  Initial check-in.
This is a replacement for Io/TextInputAdapter, with notable differences:

* It doesn't store the raw input stream, therefore neither needing to be a
template or having complications when the stream is Owned.  (TextInputAdapter
uses traits specializations to avoid accidentally losing an Owned to io::read(),
and if it's missing one it compiles anyway but misbehaves at runtime.)

* It keeps a statically allocated buffer instead of a list of strings.

* It doesn't process an entire block at a time, but only as much as is needed,
so you pay as you go instead of upfront -- a big win if there are early errors.

* It exposes its buffer, so the client can pass it directly to read() instead of
needing its own buffer which then must be copied.

* The client can explicitly request a complete line only, or any non-empty line
fragment (which does return a complete line if available).

* The line is returned as a reference.  The client may make a copy if desired.

* It's more easily debugged and stepped, and results in smaller executable size.

* It comes with unit tests.