Add make-traps.pl, which reads CIncludes/Traps.h and outputs a table.
The output is multiple records of the form:  <trap-word> TAB <trap-name> LF
where trap-word is the word in hexidecimal with no prefix (e.g. "A9F4") and the
trap-name includes the leading underscore.

Traps are listed here in the same order as in Traps.h so that information is not
lost to Perl's unstable sort.  Some trap names are duplicates; of these, some
(_InitDogCow, _EnableDogCow, _DisableDogCow, _Moof, and _HMSPinaforeDispatch)
are considered 'silly' and not included.  The remainder are disabbreviations
(e.g. _DisposeHandle vs. _DisposHandle) and follow (rather than precede) their
shorter cousins -- and quite conveniently, since they're preferred.[1]

[1] It's convenient in that we can process the file in order and use whatever
mapping occurs last.