Replace FSTree_Sequence< DynamicGroup_Details > with FSTree_DynamicGroup.
As of this change, we avoid the general overhead of FSTree_Sequence, and in
addition, several operations are no longer templated.  Lookup_Child() and
IterateIntoCache() are defined in the base class using a virtual ItemSequence(),
and another virtual function returns a pointer to the GetDynamicElementByID()
specialization needed for the Open() method of FSTree_Dynamic_N, which is no
longer a template.