Genie:  Don't select text before deleting it on Command-Delete after the prompt.
To avoid Command-Delete's usual effect of deleting back to the beginning of the
line (which would remove the prompt) we've been setting the selection to start
at the start of input (i.e. immediately following the prompt).  Continue to do
this, but just update the field in the TextEdit record instead of calling
TESetSelect(), which also draws the selection highlight (resulting in a flashing
effect sometimes as the highlight is drawn and immediately removed).