Cancel focus when uninstalling a view.
This is critical since the focus is stored as a raw pointer.  It's only needed
for focusable views (fields), but doing it once here avoids the issue of having
to remember to override Uninstall() (and remembering to call the base member)
and also makes for less code in both the source and the executable.  Also, it
won't affect performance in the slightest.