Check focus (instead of assuming) on TextEdit activate/deactivate.
Don't TEActivate() unless the field is focused.
Don't TEDeactivate() unless the field is active (saving trap calls if inactive).

If the focus is NULL, then *do* assume the field is focused.  Since windows with
multiple fields will use a focus controller to set the focus, then a null focus
indicates the absence of a focus controller and therefore only a single field.