Add AdvanceFocus() to View and override for SuperView, Stack, and TextEdit.
AdvanceFocus() takes a View pointer and direction as input and returns a View
pointer.  Either pointer may be NULL.

The input pointer is either the currently focused view, or NULL.  NULL indicates
that no view is focused and the first one available should be returned.  (When
traversing backward, the last one is considered the first.)

The direction is false for forward (the default) and true for backward.

The output pointer given a non-null input pointer falls into one of three cases:

* The output pointer equals the input pointer if the view doesn't exist in the
view hierarchy.  Currently this can't happen at top level in an active window,
since View::Uninstall() will cancel focus if the view has it, and active windows
use the global focus.  But an inactive window could have a field removed, and
upon activation the focus controller's AutoFocus() would get this result, at
which point it tries again with a NULL input pointer.  However, this happens
normally when examining a subtree of the view hierarchy that doesn't contain the
focused view.

* The output pointer is NULL if the focused view is the only focusable view.

* The output pointer points to the next (or prior, if going backward) field (and
therefore is non-null and different than the input pointer) when one exists.

For a NULL input pointer, the output pointer either points to the first (or if
backward, last) field or is NULL if no focusable fields exist.

These semantics hold true at whatever level they're applied.  Consequently, a
stack can observe that its first subview doesn't contain (and isn't) the focused
field, its second subview does (or is) but doesn't contain any additional
focusable views, and its third subview has or is another field, which it finally
returns to its parent.  The parent would see that the stack contains the focused
field and also a subsequent field.

focus ->
{
	focus -> {...} -> focus
	focus -> {...} -> NULL
	NULL  -> {...} -> next
}
-> next

AdvanceFocus() doesn't wrap around, since that can only be done at top level.