Genie [FIX]:  Don't truncate regular files after each write.
Check for both O_TRUNC and O_LAZY in RegularFileHandle::Write(), rather than one
(or both).  The effect of this logic error was to truncate after each write when
O_TRUNC was specified (which is fairly common).

One victim of this in particular was d68k, which (currently) makes very frequent
writes.