Remove the redundant search paths for CWANSIIncludes and CIncludes.
Back in the day, we appended the parent directory of the unit being compiled to
the include search path since OpenSSL needed this.  But as a result, the search
for Mac OS headers while compiling Nitrogen found Nitrogen's headers instead.
So we had to search CIncludes before the unit's directory, and since CIncludes
contains MPW's libc headers as well as the Mac Universal Interfaces (oops) then
it was necessary to search CWANSIIncludes before CIncludes.

Now, the unit's directory is not searched if 'sources' config directives exist,
which is true for Nitrogen, so the extra search entries aren't needed.