Nitrogen:  Remove CIncludes prerequisite which is no longer needed.
Previously we were searching the source directory before the system-supplied
directories (such as CIncludes), such that a search for a Mac OS header would
yield the Nitrogen header instead, requiring CIncludes to be explicitly searched
in advance.  Now A-line will skip searching the source directory if a 'sources'
directive exists, so there's no longer an ambiguity.