Factor the 68K recovery handler out of the exception handler.
The problem here is that we have to push the PC onto the stack as if it were a
return address.  The caller's stack pointer is either the USP (if the caller was
in user mode, as is the case with virtual memory or the Mac OS 68K emulator), or
the SSP plus the size of the exception frame, and we haven't even considered the
secondary SSP, the master stack pointer (MSP).

By delaying processing until after returning from the exception, we don't have
to worry about which stack was in use (though easy enough to check) or having to
shift up the exception frame to make room for a new return address (kinkier than
I'm comfortable with).