Rewrite 68K exception handler installation.
InstallExceptionHandlers() is called once to enable hardware exception handling.
Unfortunately, while this works well on PowerPC, there's no Exception Manager in
68K Mac OS.  Since the OS doesn't allow us to register an exception handler that
it calls automatically (via multiplexed dispatch), instead we'll have to install
our exception handlers in the system vector table temporarily and remove them in
between context switches.

Instead of using a table of substitute exception handlers which all happen to be
the same function, GenericExceptionHandler() and swapping with the system vector
table, explicitly install GenericExceptionHandler() for each exception we handle
after saving the original exception handlers in our table, which is performed in install_68k_exception_handlers().  We restore the original exception handlers in remove_68k_exception_handlers().

The table of recovery handlers (gExceptionUserHandlerTable) is now used (instead
of the exception vector table) as the basis (via null-check) for whether to hook
a particular vector.