Add GetNextEvent_Patch to remove exception handlers when running out-of-context.
Our 68K exception handlers must be installed while any of our user processes are
running, and must not be installed when any other application is the current Mac
process.

The _GetNextEvent patch temporarily swaps the original exception vectors back in
during the call to the upstream GetNextEvent.

This is in fact the most reliable (and therefore the cleanest) way to do things,
since it covers not only WaitNextEvent(), but other GetNextEvent() calls made by
the Toolbox.

(Such calls include OSAExecuteEvent() (either without a send proc, or when using
'display dialog').  We frown on such things, but since we can't prevent them, we
must take appropriate measures.)