[FIX] Genie:  Reset cleanup handler after running in Exec(), avoiding a crash.
Microtools don't set a cleanup handler.  The cleanup handler is initialized to
NULL in the Process constructor, but if a full tool sets one and later execs a
microtool without vforking, failing to reset it would result in a bogus attempt
to call it.

The symptom was a crash when closing a report window after invoking a diagnostic
(by selecting the Line and File commands and hitting Enter).

This wasn't an issue until recently, since previously 'idle' wasn't a microtool.