Genie:  Narrow the guard against kPrivateCFragCopy from Carbon to just OS X.
Mac OS X's implementation of Code Fragment Manager functions which can be passed
the option kPrivateCFragCopy is broken.[1]  This has been confirmed on version
10.2.8 (and it's never been fixed).  Since no later version works correctly, and
no earlier version shall be deliberately supported by Lamp,[2] we consider it to
be generally broken on Mac OS X.

This is a big performance win for the Carbon build running on OS 9, since it
doesn't keep rereading the same binaries off the disk.  The binary caching code
is unused bulk in OS X, but that's okay since there's plenty of RAM to spend on
multiple resident copies of the same program anyway.

[1] Not only did Apple decline to actually implement the documented behavior in
OS X, and not only did they also neglect to mention this in any documentation
whatsoever, the affected routines don't even return an error.  Instead they
report success, while silently pretending that you passed some other option.
Apple DTS says OS X behaves as intended in this regard.  This seems to imply
that they've intended it to be broken.  I find this saddening and disturbing.

[2] 10.2 is arguably the first viable version of OS X.  In any case, it's the
earliest for which I have install discs.  Even if a prior update of 10.2 works,
it can only matter if one doesn't install system updates.