Genie:  Override CreateDirectory() for FSTree_ResFileDir.
Essentially, call FSpCreateResFile().  The creator and type are reused from the
existing file, unless it's zero-length and has a dotless filename.  Otherwise,
'RSED' and 'rsrc' (ResEdit-owned resource file) are used.  This means that a
file whose name ends in ".rsrc" will keep its existing signature, which in Genie
defaults to 'RSED'/'rsrc' anyway.

The code is meant to handle the case of the file not existing, though currently
it will not be reached in that case since the lookup of 'res' is only special
for files, which a priori already exist.