Pedestal:  Return the code from HandleMenuItem() instead of calling the handler.
Replace MenuBar::ProcessMenuItem() with free function HandleMenuItem().  The
latter does not pass on the code to a handler (and therefore doesn't need to
store a reference to one) but returns the code as is, or zero if the item has no
associated code -- including the case of Apple menu items (i.e. system-supplied
items, not counting "About Foo..."), after it calls OpenDeskAcc() (in non-Carbon
builds).

The task of unhighlighting the menu is moved to Application::HandleMenuChoice().