iota:  Add parse_unsigned_decimal() and parse_decimal().
In MSL at least, atoi() is implemented in terms of strtol(), which is overkill
since we're only interested in the decimal case and don't need an algorithm that
works for any radix.

The new functions offer the simplicity of atoi() with the iteration semantics of
strtol() (as well as the distinction between signed and unsigned values).