Reimplement vfork() as an extern function.
Delete the vfork.h header defining vfork() as an inline function wrapping
sigsetjmp().

Remove the vfork.h include from unistd.h and uncomment the vfork() declaration.

Add new lamp/vfork.h defining struct _vfork_pad, which replaces sigjmp_buf for
this purpose.  The pad struct contains non-volatile registers (excepting A5 in
the 68K case, which is constant throughout Genie's execution) as well as the
return address (upon the call to vfork()), a slot for returning the child pid,
and a backlink (which is the previous frame pointer) in the PowerPC case, for
compatibility with alloca().  Also defined is _resume_handler_t, replacing
long_jump_t.

Rewrite vfork.cc to add vfork() as an extern asm function for each of 68K and
PowerPC.  They save the return address and non-volatile registers in a global
pad struct and jump to _vfork() (formerly _vfork_start()), which now passes to
the kernel the address of resume_vfork() (replacing kerosene_long_jump()), which
again now calls another asm function _resume_vfork() to restore saved registers
and jump to the saved return address.

Meanwhile, in the kernel, update LongJumper generally, and specifically store a
copy of the vfork pad, since the one passed is global to the caller's process
image and will be clobbered in the event of a nested vfork.  This leaves a small
window of death where a client's main thread and another spawned via reexec()
call vfork() one after another and the first of them switches tasks on entry to
vfork_start().  Since no existing programs call vfork() from secondary threads,
this race condition is acceptable temporarily.

Also, assign the pid to a field in the vfork pad instead of passing it directly
to the resume handler.