Allocate the vfork pad on the stack in vfork().
Remove global_pad and its associated race condition.  In vfork(), allocate stack
space to store the vfork pad, and pass its address to _vfork().

For 68K, we have to push A7 and A6 separately since we've already modified their
values by the time the pad is allocated.  On the other hand, we now push D0, not
because we care about its value (which we ignore), but because it decrements the
stack pointer an extra longword for free.  Finally, we push the pad address, JSR
to _vfork(), and clean up.

The PowerPC case is fairly straightforward, though we have to copy the SP before
allocating a stack frame.