Narrow the system call number to 16 bits within the kernel on 68K.
This allows us to replace MOVE.L with MOVE.W in syscall definitions without the
need for a longword extension.  It also limits us to 65536 system calls[1], but
that should be enough for anybody.

[1] Number 0 is valid.  Normally the last valid system call number is followed
by one more (to which all invalid numbers are mapped) but if all possible values
are valid numbers, then we don't need the catch-all.