Genie:  Synchronize the selection when jumping to end of buffer on key-down.
In Console_KeyDown(), we select the end of the buffer on Return or if the input
area doesn't contain the selection.  Previously we were calling TESetSelect()
only, with the result that subsequent output would restore the old selection.
The solution is to update the data model's selection at the same time as the
TextEdit record.