MSL-fixup:  Fix the conversion from MSL __io_modes to POSIX mode_t.
Since __append is a flag and not its own mode, check for it as such.

Previously, .git/FETCH_HEAD was opened with mode 0x0200 (O_RDONLY|O_CREAT) and a
subsequent FSWrite() call would throw wrPermErr (OSStatus -61) at the end of git
fetch, with results ranging from none at all to an uncaught exception (depending
on which recent patches were active) but in all cases leaving the file empty.