Nitrogen:  Return an owned struct from SPBGetIndexedDevice().
Previously we returned a struct containing an owned member, which is all kinds
of painful.  Instead, let the entire struct be owned.

The SPBGetIndexedDevice_Result struct is now named SPB_IndexedDevice, and the
former is a type alias for owned< SPB_IndexedDevice > (which is now the return
type of SPBGetIndexedDevice()).