Nitrogen:  Replace SoundInputDevices() with SoundInputDevice_Names().
Also, remove the compile tests which no longer apply.

SoundInputDevices() is a sequence based on SPBGetIndexedDevice(), which returns
an icon handle in addition to a name and whose result therefore must be owned.
Unfortunately, this wreaks havoc with IndexUntilFailureContainer (which must be
used since we don't know the end of the sequence except by getting an error
traversing it), since its iterators must store a copy of the value (unless we're
willing to call the value-getting funciton on increment *and* on dereference),
making the iterator uncopyable unless we add Transfer machinery.  This wouldn't
be so bad if it weren't the generic code getting janked up for the benefit of
one poorly designed client.

SoundInputDevice_Names() is a variant based on SPBGetIndexedDevice_Name(), which
just returns each device name as a Str255.  The only user so far is Genie's
/sys/mac/soundin FSTree subclass, which doesn't even need that much -- all it
does (with the sequence) is count the number of elements (since the directory is
populated by the indices, not the device names).