[FIX] Rewrite array_singleton for compatibility with g++.
The previous array_singleton was written as a class template parameterized only
on the class of elements in the array, and the type member array_type was (at
least by intention) a class template with eight parameters of the element type
(all defaulting to zero) and derived from array_singleton_8, but array_type was
also specialized with fewer parameters to derive from the shorter array classes.

Although this works in Metrowerks C++, it doesn't in g++ 4.0.1.  Specializations
alter implementation but not interface -- that is, foo< 0 > might behave totally
differently than other foo< n >, but the template< int x > interface can't be
specialized away -- it must be instantiated with a template signature matching
the original declaration.  The previous attempt to work around this limitation
using default parameters resulted in substitution of the defaults (as desired)
but never matched a specialization in g++, so every array singleton had eight
items.  This broke CarbonEvents except in the case of event handlers accepting
exactly eight arguments (not counting user data).

It also had two warts:  The zero-element case didn't work in CodeWarrior 8.3
(requiring a workaround in CarbonEvents.hh) and the resulting type is written as
array_singleton< Element >::array_type< e0, e1, e2 ... >.

The new solution defines array_singleton_n, a class template whose parameters
are an integer number n of significant elements, the element type, and eight
element values.  It's specialized for different values of n to derive from the
matching array_singleton_N class (where N == n).  The class template array_size
defines a static int member as the number of significant elements (those which
are not trailing null values).  Finally, the new array_singleton derives from
array_singleton_n, which n equal to the array size.

The zero-element workarounds are no longer applicable or necessary, and usage is
now:  array_singleton< Element, e0, e1, e2 ... >.