Nitrogen:  Just try for actual OSStatus in ConvertTheExceptionToOSStatus().
Rethrow and try to catch as OSStatus, or just return the default.

TheExceptionBeingHandled is clever, but currently no use case exists to justify
the cost incurred by every user of Apple events or Carbon Events.  One possible
future option is calling a user-settable routine, as with terminate(), allowing
the application to plug in the exception conversion machinery.