glibc:  Define REGEX_ALLOCATE() as function-style to avoid issues with jcpp.
Since jcpp can't handle expanding a function-style macro (alloca) when the name
is the result of expanding another macro (REGEX_ALLOCATE) without completely
rewriting the parser, avoid the issue by making REGEX_ALLOCATE() include the
size argument.

A parser rewrite is due in any case, but this patch allows us to postpone it.