gzip:  Constify the char* argument to error() to appease Metrowerks C/C++.
While not an issue when MWC compiles util.c, given preprocessor output util.cpp
(which the compiler assumes to be C++) beginning with #pragma cplusplus off, MWC
reports passing a string constant to error() as an illegal implicit conversion
from const char* to char*.

It appears that the pragma is not switching off all C++ behavior.  The easy fix
is to make the code compatible with both languages.