extended-api-set:  Implement renameat().
Previously, our strategy has been to fchdir to the dirfd, make the old-style
system call, and fchdir back.  We can't do that with rename() since it has *two*
dir fds.  The new strategy is to call realpath() as needed (temporarily setting
the current directory) and finally calling through to the system call.