A-line:  Don't define TARGET_API_MAC_{CARBON,OS8} for Blue -- the defaults work.
It's only necessary to set TARGET_API_MAC_CARBON to 0 (and consequently also set
TARGET_API_MAC_OS8 to 1) prior to including Carbon.h, since otherwise it assumes
we're targeting Carbon.  But (a) we don't include Carbon.h without including
ConditionalMacros.h first, and (b) even if we did, then that would be the point
at which to define these macros.