plus::string:  Fix conceptual error -- initialize the margin to the max offset.
If we (a) allow the string to store NUL bytes, and (b) store the length (as the
number of bytes unused) in the last byte, coinciding with the trailing NUL of a
string filling the entire space, then we can't (c) zero-initialize to a default
state.  While elegant, this is not important, and in practice strings won't be
default-constructed statically nearly as often as in vectors (i.e. non-static,
uninitialized memory).

Therefore, a margin of zero always indicates a maximized small string, while an
empty string has a margin equal to the total number of bytes available for data,
which is also the offset to the margin byte.