plus::string:  Factor private reallocate( size_type ) out of assign().
The reallocate() method includes the guts of assign( const char, size_type ) for
the purpose of managing a reallocation, but without actually copying any data.
It exists so assign( const char*, size_type ) and assign( size_type, char ) can
share common code.

Since copying the data now happens after (a) the terminator byte is zeroed, and
(b) the old buffer (if any) is disposed, two failure modes arise:  One, if we're
assigning a subset of our existing small-string buffer, the input data could be
clobbered by the zero byte.  Two, if the input source is similarly an allocated
buffer of ours, it's invalidated by dispose().  To solve the first problem, we
specifically check for the input overlapping the small-string buffer, and if so
we construct a new temporary string with the input instead -- which can't suffer
the same issue since it occupies uninitialized stack memory -- and then swap.[1]
To avoid the second issue, we swap with a temporary before calling reallocate(),
so the old buffer isn't disposed until after the copy has occurred (whether it's
the data source or not).

[1] There's no valid case where input overlapping the small-string buffer is not
also wholly contained within the buffer, but overlap is what we check for.